Add the Rat Emulation Machine
As I identified in the last commit, we had to axe the Immobile Killer
Machine because we had no way of sensing and/or identifying targets as
they approached us.

I had deliberately removed movement from that machine to try and make
attack development easier. Turns out, it was gonna be easier to
develop a machine that managed movement then one that ignored it. We
needed a really basic machine, that fled threats and and moved towards
desirable objects/goals.

To me, that sounded like a rat! Thus, I've created the Rat Emulation
Machine and the associated Pawn and Test scene. It's pretty rough, but
right now it just focuses on fleeing from static bodies it detects.
It's gonna take a lot more work to get this working - let alone
documenting everything - but this is a great start for now!