Add Rat Emulation graphs (and README updates)
The last thing we needed to add to complete our documentation push was
the graphs for the Rat Emulation Machine.

The heirarchy was easy enough to model, but we needed to split apart
the flow graph into two separate graphs: the control flow (which is
the standard) and the brand new data flow graph. I felt that we needed
to show how variables were being manipulated and what was listening to
which signals - the REM is just SO complicated that anything less
would have felt insufficient.