Move error tracking in the REM into Walk substate
We had our error tracking variable, _targ_dist_history, sitting on the
Physics Travel Region where it could be accessed by anything in the
Rat Emulation Machine.

However, while developing the documentation, I realized this wasn't
necessary. The only state that required the variable was the Walk
substate. It would be much cleaner to have it just sit on the Walk
state where it was actually needed.

So, I did that. A minor change but one that I'm definitely happy with.