Add a Pathing Interface Core
Our navigation nodes - the vanilla Navigation node and Milos Lukic's
DynamicNavigationMesh node - must, by their nature, exist outside of
any body that's going to be moving around and referencing that
navigation node. Otherwise the whole navigation mesh would move with
it!

This was the reason navigation and pathing was handled at the
integrating body level. The KinematicDriver was designed with this in
mind - we would have position reference function configurable that
allowed each body to specify it's own position, and would be handed
positions by some external force.

However, with the RatEmulationMachine, we've run into a problem. The
individual Goal States need to set paths for the PhysicsTravelRegion,
but have no way of doing so. We needed a way to pass the navigation
data down to the Goal States. We needed a new core.

That's where the PathingInterfaceCore comes in. This core holds the
navigation resources and provides functions for generating paths with
those resources. It does still require some runtime configuration, but
this solution is more elegant and more useful than tacking that onto
the KinematicCore.

The KinematicDriverMachine and KinematicCore now require a makeover,
or perhaps a simplification. As do all the associated tests and test
assets...