Retrofit Kinematics to use Pathing Interface Core
With the introduction of the Pathing Interface Core, some elements of
the "Kinematics" (the Kinematic Core, the Kinematic integrated bodies,
etc.) were now moot.

The Drive Body and Position Function configurables were now pointless.
Indeed, the concept of handing fully formed paths to the integrating
bodies and several scenes was now forfeit. So too was the concept of
"choosing" a method to determine a body's current position.

This necessitated a lot of changes. The Slope Step Test, suceeded by
the Advanced Kinematics Test, has been wholly removed. Those bodies
that integrate the Kinematic Driver Machine (as well as the machine
itself) have a new function: move_to_point. Provided a Vector3, the
body will path to that point (if it can). In cases where made overtly
complex paths (such as the Viewport Shader Test), this has made the
code a bit... uglier. However, in other places (like the Advanced
Kinematics Test and the Dynamic Navigation Mesh Test), it's made code
a lot better.

I also patched up some errors along the way - though they were to
numerous to count here.