Stop REM state-lock on no path
I identified a potential issue with the Rat Emulation Machine - we
relied purely on signals coming from the Pawn to leave the state.
However, if we weren't able to get a path when we entered one of the
REM's Goal states, then we can't fire a signal, and we're effectively
locked in a state.

So, I added checks in the Flee and Wander states. Once we set the
path, we double check that there is actually pathing data. If there
isn't, then something went wrong. We then go back to the idle state,
saving ourselves from potential state-lock.

Also, to make transitions more explicit, I added a "GoalRegion/"
identifier on the Idle transition call.