Update the physics layers
As I was working on the Rat Pawn's test level, it was becoming
apparent that we needed to adjust the physics layers. Pawns didn't
need their own layer, we needed layers to denote hazards, etc.

So, I changed up the layers and updated the documentation. Something
new I cooked up is the concept of "blocks" - the different physics
layers are organized into these blocks so that we have a better idea
of what's happening.

I also changed how we do certain layers, especially the terrain.
Rather than having two terrain layers, one raycast blocking and one
not, we have one "Terrain" layer and then two layers to give
attributes to that layer, such as "Raycast Blocking" or "Camera
Obstruction".