Use three different areas for the Sensory Sort Core
Previously, we had just one sort of general sense-everything sensory
area. The problem with this was one of priority - we don't want to
flee at the exact same time we detect a target.

I felt like it'd be good to have multiple layers. Nested areas, with
different priorities. This does some of the heavy lifting for the
machine nodes, making their internal deliberations easier.