Sort bodies via Group with a GroupSortMatrix
We needed a way to identify different spatial elements in this game:
hazards, items, projectiles, Motion AI, etc. It had to be at-a-glance,
but also be robust. We did some of that with the physics layers, but
there are only 20 layers. We had to be very sparing with how we
delegated those out. We needed a better solution.

Though poorly documented, Godot's "Groups" system was the closest thing
we could get. It's kind like to a meta-tagging system. Every Godot Node
can be a member of a group, so checking membership in the group allows
us to treat all sorts of different objects in a consistent manner.

The SensorySortCore already provided us with a methodology for sensing
items - now we needed to do the sorting of the sensed objects. So we
had to do that with groups.

However, I decided to build a dynamic system for interpreting these
group tags. So, rather than hard-coding that Team 1 is an enemy of
Team 2, each entity of the team would have it's own system for
devising which group-tags constituted "enemies".

That's where the GroupSortMatrix comes in. It provides several hard
coded "group categories" (just two, for right now). This gives us
consistent classes for the Machines to react to. What's really
promising is that we can have as many or as few group strings in each
group category. So we can have NPCs that don't see enemies, scardey
cats that sense danger, fools that hoard wealth. All we have to do is
add or remove the tags to change our behavior!

The GroupSortMatrix has been integrated into the SensorySortCore, as
well as the RatEmulationMachine. It required a lot of changes, but I'm
happy to say that everything seems to be running and working really
well!