Add in a flee animation
I decided to create a special form of the walk animation just for
pawns that were fleeing. The result was... decidedly overwhelming. I
think the Pawn and Pawn-type entities are the only things that are
going to end up having "true" animations. I really just cannot STAND
animating things, it's not where my interest lies. I think that, in
the future, we'll make use of Godot's animation structure. Maybe.