Add in Qodot and try it out
A big gripe that I've had with Godot has been the difficulty with
creating world geometry - I spent about a month making procedural
building components just to get around doing it in Blender.

But something's different now - now there's Qodot. Qodot allows us to
import Quake style .map files into Godot as the appropriate entities:
physics bodies, areas, etc. Combined with the modern level editor
TrenchBroom, this allows to quickly and easily design complex
geometry. It's quicker then Blender, it's easier then Blender and,
because I used to make L4D2 levels, it is very easy for me to use.

However, converting to use Qodot and TrenchBroom did give require a
pretty major rework: we were required to place any texture we wanted
to use in TrenchBroom in a "textures" directory. This meant we had to
migrate most of our textures.

To try it out, I started by more-or-less recreating the
RatEmulationTest. Seems to be working so far!