Update the Qodot definition files
Qodot came with a whole host of different entities and the like for
integration with TrenchBroom. Only problem was, I had no use for them.
There was a lot of them, and some of them weren't set up in a way that
I liked nor cared for.

So, I deleted them and then added my own. It took a lot of trial and
error but I eventually settled on a group I was pleased with. I'm sure
I'll get more as time goes on. I should probably document these at
some point...