Update Breadcrumb class
Previously, the Breadcrumb class was called in "patrol_breadcrumb".
However, this caused an issue - Trenchbroom grouped the point class
under the "Patrol" category as just "breadcrumb".

Except "Patrol" wasn't really a category. It was just a name. So,
after thinking it through, I decided to rename it to "ai_breadcrumb",
since an "Ai" category we'll probably use later.