Add a haul animation
Since we're gonna be adding moving items around into the game, we're
gonna need an animation for the Pawn moving things around.

That's what this new "Haul" animation is for. My plan is that the
items will hover a fixed distance above the Pawn's head, giving the
items the appearance that they're being held over the Pawn's head. I
THINK that will look good. I HOPE it looks good.