Add an interaction range to the SensorySortCore
As I developed a system to grab items - one important question kept
popping up: how do we know WHEN we can interact with an object?
Initially I was working with the idea that objects would have their
own areas attached and would notify passing bodies that they were in
a space to pick up an item.

However, the more I thought about this, the more I realized that items
should be items. They should be self contained. It is the
responsibility of the agent to determine what it can interact with,
not the item/interactable.

That meant we needed to expand the SensorySortCore with a brand new
sensor: the Interaction Range Area. This will allow any integrating
machines to detect different kinds of interactables at will. Now that
we can detect these interactables, we just have to do something about
them...