Crash the game by messing with physics (oops)
Now, normally, I wouldn't commit broken and code, but I had to make an
exception here - this crash underscores a pivot in how we're gonna
handle items from now on.

See, I had the bright idea of having the Pawn actually GRAB physics
items by changing the collision layers and mask and all that nonsense.
That way, we wouldn't need to constantly spawn and destroy items.

Well, I'm glad to report (if the name of the commit didn't give it
away) that THAT idea was TERRIBLE. Changing the collision of an item
causes it to retrigger the "enter" and "exit" signals on any and all
areas. That's actually desirable behavior, but since the
Packing-Sorting-Rat Machine I was working on had basically hardlined
that functionality into the changing-the-physics part of the code, I
created an INCREDIBLE and TERRIBLE stack overflow error everytime I
ran the basic item test.

The whole idea behind the different item states was so that we didn't
need to differentiate between an item's physical form and it's form
when being carried or otherwise stored. We definetly won't be able to
cut that any longer. We're gonna need to separate an item into it's
purely physical appearance and the physics body part. Surely there
must be a way we can do this all as one scene via the MasterItem
scene. It's really gonna bug me if there isn't...