Create Items that can be physical or visual
Okay, so I wanted to add items to the game, but was presented with an
interesting problem: how do we disable or enable it's "itemness"? Like
when an object is mounted on a wall, or stowed in a crate?

The solution I ultimately decided on was one of shells and components.
There's a MasterItem class, from which all items are derived from.
There are two key child nodes - the VisualComponent (which contains
the visuals) and the DataComponent (which contains the data). These
are repeatedly passed back and forth between the MasterItem-inheriting
class and the VisualItemShell, which exists JUST to house the
components. In this way, the critical data gets preserved but the
unecessary chuff gets dropped.

It seems to work really well with our Pawn grabbing at stuff. There's
still a lot more work to do, like figuring out how to actually drop
stuff. But I'm happy for now!