Design a basic task-action system
We've been doing pretty well with the machines and systems up until
now. However, as I was developing a better test for picking up an item
and placing it down, I realized that how I was doing it kinda...
sucked.

I was trying to do everything in either the goal states, or substates
there-of. What I quickly realized was that there was no way we could
do these things in states that are integrated into the different
machines. We needed a more acceptable "line unit" that we could swap
in or out.

That's the idea behind the tasks & actions. The actions are (mostly)
atomic, pass-or-fail actions. Like, "Go to Point" or "Grab Item". They
are implemented as individual XSM states. These actions are then
strung together and managed using Tasks - which are also XSM states.
The Tasks handle the feedback from the Actions and handle the
transitions between them.

Eventually, tasks will be created and destroyed on the fly as the
machine demands. For now, I've just designed one task - MoveItemDrop.
That really translates to "Move an item by picking it up and dropping
it again".