Begin integrating task-actions into an AI Machine
We developed the Task-Action paradigm in previous commits - the next
step was crafting a system that allowed us to swap tasks in-and-out at
will as part of the machine's typical functions.

Thus, the Packing-Sorting Rat Machine has become the Tasking Coward
Machine - because it does tasks but it also still flees. There are
several key differences between the Tasking Coward machine and the
machines that preceeded it.

First, the machine as a whole is no longer a singular XSM state
machine. It is composed of anywhere from two to three state machines
at any given time. Although having the whole AI Machine as one
singular XSM Machine was very beautiful and offered a lot of utility
(beautility?), there was a problem: XSM states did not support
dynamically adding or removing child states. And since I had decided
to model tasks as mini-machines, that meant we couldn't organically
inject and remove the tasks without messing up the whole XSM machine.

So, I figured there were two choices: either change the AI machine to
allow for injecting the micro-task machines, or change the tasks and
actions so that they were no longer states. XSM just offered such a
natural way of structuring the actions and chaining them together as
tasks that I ultimately wasn't willing to change how tasks or actions
worked. Besides, the more I though about it, the more I realized that
a unified XSM-AI Machine wasn't necessary: since we had the regions
always running, we (essentially) always had multiple mini-machines.

Once the XSM machine had been retrofitted into collection of machines,
I added the brand-new TaskManagerRegion. The TaskManager will be the
key authority for adding and removing tasks. There can only ever be
one task in play, and tasks (should) be removed when entering and
exiting a scene.

This creates a three component paradigm for our AI machines: There is
the high-level goal, the low-level tasking and actions, and the
physics processing. Hopefully that's enough to handle any AI needs.
Hopefully...

Finally, in order to make sure this was working, I eliminated the
"Wander" state and recreated it as a task that is occassionally
scheduled by the Idle state. Everything looks good so far!