Create a Flee Task for the TCM's Flee state
Wandering and idling had already been handled via tasks - now was the
time for the fleeing to be handled with tasks.

So, I added a new action - FleeAllBody - and a task to wrap it, the
FleeImmediateThreat task.

There was a great amount of consideration that went into making
"fleeing" it's own task. I wanted to just use the MovePosition state
since "fleeing" is technically just moving to a position that keeps
changing. However, the semantics of moving AWAY from something (or
multiple somethings) was different enough from moving TO something
that it got real complicated real quickly.

I imagine "FleeAllBody" will eventually find a home with a whole bunch
of other "Flee" actions that offer different utility.

Already, I'm starting to see an interesting dynamic crop up - where
should complex behavior lie? With the Goal State? With the Task? With
the actions? If you looked at this commit, you'd think I'm leaning
towards complex actions, but I'm leaning mostly towards complex tasks.
That's where the real complex back-and forth should be. I think. We'll
see as the project goes on.

For right now, I'm testing the Tasking-Coward Pawn in the Rat
Emulation test - just because that Pac-Man Rat Maze is so useful!
We will need to develop a better test, specific to the Tasking
Machine - eventually!

Also, renamed an action and several references to a "Range Area". I'm
not sure what I was thinking when I put that in - we already have the
Priority Area paradigm.