Update the Pawn's collision mask
While testing the Tasking Coward Pawn, I noticed it was sinking into
the ground for some reason. I recognized this as a collision issue,
and I know I've seen it before, but I couldn't remember how I fixed
it.

Turns out, it was the collision mask? Colliding with the Hazard layer
was causing the issue. The Hazard layer is supposed to be immaterial
or intangible anyway, so I guess it doesn't matter, but still, I'm not
sure exactly why that was happening.

I decided to dedicate a commit just for this fix - just in case it
ever happens again...