Actually move an item via the tasking system!
Okay, I've been a bit distracted lately but I did get to do some
thinking: I had originally wanted the TaskingCowardPawn to be able to
do ANYTHING we tasked it with, but the more I've thought about this,
the more I realized it wasn't viable (at best). The real trick is that
I designed tasks such that they need to be initialized by the machine
that's using them. It makes this sort of artifical insertion a bit
difficult.

That had reall jammed my progress for a fair bit. However, after
thinking it through, I decided that it was important to just get the
capability in rather than getting it right the first time.

So, after a little bit of working, I've managed to actually move an
item via the tasking system! It's a bit simple, but for right now,
that's all it needs to be. The conviction that it should be more
complicated was holding me back.

It took some finangling - I had to update the Move Item task a bit, as
well as the XSM stuff, and add a new goal state. But for now,
everything looks good. But where to go from here?