Drop physical items in same location as visual shell
Previously, when we "dropped" an item, it was creating a new item
instance at the world origin. It was doing that regardless of whether
or not the item was AT the origin.

The fix was simple enough - we just needed to feed the Visual Shell's
global transform to the new Physical Item when we "physicalize". Now
the transition between physical item and visual item is almost
seamless.