Convert the PathingInterfaceCore to a "Level" Core
Previously, the PathingInterfaceCore was our one-stop shop for
everything and anything related to pathing.

However, as development proceeded, I realized that a pathing-specific
core was a little bit TOO specific. I had other needs that didn't have
a good place, but related to the level-at-large. For example, where do
we put loose items so they are not freely scattered across the scene
tree?

So, I recast the PathingInterfaceCore as the LevelInterfaceCore. This
will hopefully serve as our general interface for level-specific
information.