Use a dedicated item parent, via the Level Interface
Previously, when we dropped items, we made a vague guess about where
the items should go. I didn't like that idea because it was stupid.
Seriously, it's like throwing a dart at the scene tree and just sort
of hoping it doesn't impale anything. It's just asking for trouble.

I figured it would be better to have a dedicated item parent. One that
could be passed down to an Item Management Core so we could have a
sort of consistent "dumping ground".

The is now the "Item Parent", occassionally also called the "Item Drop
Parent". It's stored in the Level Interface Core, and is accessed by
the Item Management Core. This means that the Item Management Core now
needs a Level Interface Core - our very first inter-core dependency.
However, since I'm not really a fan of such a sibling dependency (it
feels messy from a software perspective), it's an optional dependency.

If we don't get the Level Interface Core, we instead... do the stupid
parent-of-our-parent thing. Oh well, nothing's that perfect...