Eliminate Visual/Physical items, use RigidBody
Items need to have two states: a physical state where they can fall
over and do physics STUFF, and a purely visual state where they don't
collide with (most) anything.

Our previous solution was a bit weird - the physical and visual items
were implemented using different scenes. Transitioning between the two
meant spawning new shell scenes and moving certain nodes across.

This was originally developed in response to a stack overflow error:
we were grabbing items when we detected them, but the change in state
caused another detection, triggering another attempt to grab, casuing
anothe detection, and so on.

I was never really a fan of the system, but I was moved to change it
after I had a couple of realizations. First, that grab-in-response
code was old test code. It was always a placeholder, a prototype of
some kind. Yet I made a major design decision based on the fact that
the prototype code was broken. I also realized that, since the Tasking
system as designed separates detection and grabbing into different
Actions, it's very likely we'll never run into this detect-grab
overflow again.

With that in mind, I eliminated the Physical/Visual Item shells; these
now exist as different states of the MasterItem class. The different
functions for changing the state have all been integrated into the
MasterItem class, and supporting functions have been changed thusly.

This is now... much nicer. Just nicer to look at. I think there may be
a chance of hitting the detect-grab overflow again, but only in that I
can't PROVE that it can't happen.