Implement a hauling animation for the TCM-Pawn
I created a "hauling" animation a while ago. However, I'd never really
had any purpose for using the hauling animation - yet. I made the
animation in anticipation of hauling, but hadn't actually implemented
it.

Until now! I added in the hauling animation - both idle and walking
forms - to the pawn frames and also added in a utility function for
seeing whether the Item Management Core has an item. We could have
done just a simple check, but I wanted the explicitness that comes
with having a dedicated function. Reduces the need for duplicate
comments, ya know?