Add in 3D Box and Barrel items
I had coded in 3D items in theory, but hadn't put them in to practice
yet. I figured we'll do 3D objects eventually, just didn't know when.
But we did need SOME sort of test.

So, I created two maps in TrenchBroom - one of a barrel, the other of
a box, and then imported those into Godot. I really do love
TrenchBroom, it makes these modelling exercises a cakewalk. To texture
them, I grabbed more open source game art off opengameart.org. I'm
really grateful that it's there, though the parentage of these
resource packs can be pretty complicated.

I then slapped them in the scene, and the work pretty well so far. Not
sure what I want to do about collision exactly. The box and barrel are
large enough that it's weird when objects don't collide with them but
right now they kind of fly off the map when colliding with an NPC.
Hilarious but not exactly good for a game, is it?