Add in a standard proving ground
I was going to create another boxy level for an AdvancedItemTest, when
I realized that, if I was just going to keep making the same stupid
box over and over again, I might as well just make a nice box in
TrenchBroom.

And so I did! The nice box is a bit bigger around the edges so that we
have some more room. I stuck it in with the Advanced Item Test and
everything seems to be working just fine. Neat.