Move multiple items at one time via a Task
Okay, so previously we could only move one item at a time. I had
designed it to be that way on purpose, for simplicity's sake.

However, the more I thought about it, the more I realized this was not
going to work long term. Especially with instances where we had to
pick up lots of small items. Like apples, or cheese, or way too much
bread. We needed to be able to move items en-masse.

We may play with the approach as time goes on, but the solution I
settled on was allowing stacks of multiple items, set per-item class.
This will allow us to tune the stacking for each item. The task I
designed to retrieve the items prioritizes items by closest distance,
and includes for fail-over if an item can't be retrieved.

I ultimately ended up deleting the move-a-single-item task, since the
processing ended up being really different from the move-multiple
task, and it's easier to just move a single item by ordering the
move-multiple task on a single item. The drop item action also needed
to be modified so that the whole item stack gets dropped.

Now then - I think I need to add some obstacles to test those fail
states would go a long way...