Test item moves by including failure situations
Our previous tests for the item moving didn't really have fail states
because there was never a way for an item to be inaccessible. Well,
maybe that was true in the Basic Item Test if an item got punted off
a cliff. But I wanted a purposefully designed failure case, instead of
a merely incidental one.

So, I added some cages and breadcrumbs to the Proving Ground map. We
now use these little collectives of breadcrumbs to spawn the items and
determine where we haul them. The caged items test how we handle items
that are inaccessible - and also ensures we don't accidentally grab
such items through the wall.

I feel like we have a pretty robust item system now. We still have to
do work on containers. And there's a documentation push I've been
putting off. And then comes attacks...