Update Texture docs; Move root TrenchBroom directory
Gave the textures directory a quick passover - there were some typos
in the license and the documentation didn't address the "items"
directory.

I also moved the TrenchBroom directory that we had sitting in the root
directory down to the items/furniture directory. I'm really not sure
why I had that directory at the root to begin with. I can't really
think of a good reason and I didn't leave one when I committed it.