Add documentation for the Tasking Coward Machine
The TCM was undocumented, so I've updated the documentation. I made
the decision to not touch on task and actions too much in the machine
documentation, since that is going to be its whole own documentation
section.

I also split the control-state flow graph into two separate graphs:
one for the Physics Travel Region, and one for the Goal Region. Since
the machine is now an XSM-hybrid composite... thing, I figured we
could greatly simplify the graph by separating the two. They're not
part of the same XSM, so it makes little difference.

While developing the diagram, I added the Sensory Sort Core to the
data flow diagram. It's an important part of how the process works,
and leaving it out felt like a glaring omission. I then updated the
equivalent Rat Emulation Diagram, so that it was also on the same
page.