Change the Move-Item-Drop to allow a soft-fail
While documenting the tasks, I realized there was some ambiguity in
the success and failure of the MoveItemDropMulti task.

See, if it ran into a problem - say it got stuck - the task would
fail, whether or not it had delivered any of the items. However, if
it didn't encounter such a movement problem, it would succeed. In
fact, it would succeed EVEN if it hadn't managed to grab/deliver all
the items.

This meant our failure state was ambiguous as well as the success
state. I realized this while documenting. I like having the two
states, so I decided that the failure state should be the only
ambiguous state. If you receive a failure signal, you'll need to
evaluate the extent of the failure. If you receive a success signal,
it is guaranteed that the task succeeded!

I've updated the MoveItemDropMulti task to reflect this reality. I
checked the other tasks but none had the possibility for a soft
failure.