Document tasks and actions
I left the tasks and actions for last because they were the newest
features and also deviated from anything else we've done before.

These were a little bit more difficult because, unlike other items
we've documented, I had to make a more deliberate effort to document
the internals of the tasks and actions, and go into more intimate
detail about how they worked. After all, it's sort of a core component
of how AI works, so we had to describe it at a really basal level.

But that's okay, because it's done now! Only issue now: I've been
documenting so long I forgot what else I was doing...