Separate direct move speed refs via Kinematic Core
After a long and only semi-desired break, I'm coming back to the ol'
FicGuac project. Our next goal is to integrate attacks. However - to
be able to attack and remove health, we must first have health to
remove. To do that, we're gonna create another core for stats of some
kind.

And one of those stats is going to have to be move speed - it is a
aspect of physical being, after all. A true stat, indeed. So it was
necessary that we change how move speed was integrated. I wanted to
keep the KinematicCore's speed, so that is now the Fallback Move
Speed.

We're still using the Fallback Move Speed in all scripts, but its now
wrapped in a local variable so we can easily swap it out for whatever
core we have.