Add a CharacterStatsCore
All the different nodes up until this point have been more or less
focused outward: on how the Pawn moves through the world, how it
interfaces with that world, how it detects and organizes physics
objects into goals, threats, and desires.

What we needed now was an apparatus for tracking the state of the Pawn
itself. The state, and the stats. Since this game is going to be a
sort of chaos-RTS (or something like that), I believe that status
effects are going to play an important role in the game. Thus, it was
important to separate out the stats (i.e. HP, speed) into a core where
they could be independently manipulated.

That's the CharacterStatsCore! Right now, it has a configurable base
values for hitpoints and speed. These then translate into an effective
speed and effective hitpoints. The idea is that, eventually, various
status effects will be applied to the core and the 'effective' values
will be updated dynamically using the static base values for
calculation.