Retool tasks to allow for multistage construction
Previously, we had two initialization functions for each task: the
deliberately hidden _template_initialize() function, and the
independent initialize() functions. The independent initialize()
functions were defined seperately on each task, and each had to call
the _template_initialize() function.

This was good because, outwardly, there was only one function to worry
about - the initialize() function. But that wasn't really true, since
every function needed to call _template_initialize() in it's own
space. Plus, since all the variables needed to be provided to the
initialize() function when we called it - this meant the only place we
could create a task was exactly when we needed it embedded in the
machine.

That meant if we wanted to do something kooky - for example, creating
tasks and then feeding them into a machine from an external source -
was difficult, if not impossible. Trouble was, that kind of "task
injection" was ideal for the TaskingCowardMachine. It'd allow us to
only have one task state and just have X functions to order the
machine around.

So I changed our two functions into three functions - the generic
template_initialize() function, the task-oriented
specific_initialize() function, and the more general initialize()
function. All the task has to worry about is overriding the
specific_initialize() function. And, unless you're doing something
weird, you need to call the initialize() function.

We now use an argument dictionary to give task-specific arguments.
This meant we needed to add argument-key constants to the different
tasks. It's a bit annoying but... eh. It works.