Accept specific-initialized tasks in the TCM
Previously, the plan was to have one function on the TCM and TCM-Pawn
for each different task we used to test it. After applying one extra
second of thought, I realized that was a TERRIBLE idea and immediately
thought of a better one.

Since we separated out task initialization into specific and template
initialization, I realized we didn't need to do ANYTHING 'specific' in
the Pawn/Machine level. So I changed the move-specific function into a
more general function just for passing a task.

Now we do the task-making in the different tests, which aligns MUCH
more closely with how we'll actually be doing things. Different tests
for different functions, and all that.