Add a rudimentary proto-attack task
I'm not sure how to work on damage just yet, so I decided the
first-and-best thing I could do would be to start developing a basic
attack-task structure.

So I added a new action, MoveToEntityDistance. Like the other
MoveToEntity action, this one attempts to close distance on a
particular entity, but relies on an actual measure of distance instead
of a detection area. I also added a TimeWait action - an action all
about waiting. You just wait for a configured amount of time. That's
all.

I then bound these together into the MeleeAttackTarget task and
updated the AttackModelTest to actually test out the changes.

The result? It was bad. Actually really bad. The Pawn stuttered
constantly, continually, endlessly. I think I need to re-examine how
we do movement, which will be a really big change. But the game isn't
going to work if we're stuttering trying to follow a moving enemy!