Remove first path node to prevent cycle stutter
We had previously encountered an issue where any Motion AI would
"stutter" when attempting to re-path too frequently. We had always
sort of tip-toed around it, but with the addition of tracking a moving
target, the Motion AI had floundered and become useless. Stiff! It
couldn't move at all!

After some digging around, I found the culprit: for either "type" of
path (Dynamic Addon vs Godot Native), the first "point" or "node" in
the path-stack was ALWAYS where we started. At least it seemed to be,
I can't speak to infinite possibilities and all that guff. Anyway, the
problem was we passed off this path stack to wherever and this meant
we always lost a "cycle" to processing the point WE WERE ALREADY AT!

So I've added code to remove the initial point/node from the path
stack. That seems to have fixed it, so we'll just keep an eye on that
for now...