Disable some specific Godot warnings
Godot, like a lot of descendant C/C++ descendants, offers up several
suggestions for best practices in the form of warnings. However, I had
problems with some of these warnings.

Some of the warnings actually ran counter to my personal style of
programming. That wouldn't be a bad thing, but these conventions are
so widespread through the code that one test generated 7000 WARNINGS.
And, when that many warnings are being generated, it rendered the
warning-and-errors area utterly useless.

The first warning I disabled was the "unused argument" warnings, which
neatly insists that I append an underscore to the front of any unused
function arguments. This would be good practice... if I hadn't
accidentally filled the project with such function calls. See, much of
the project is designed as a library instead of "bespoke" solutions so
unused arguments are incredibly common. I was unfortunately unaware of
this convention until far too late, and a refactor now might very well
drive me to madness.

The second warning I disabled was the "unused signal" warnings, which
I disabled for much the same reason. The signals here are meant to be
used in an almost-library like context. Hence, an unused signal is
perfectly acceptable.

The third warning I disabled was for unused/discarded return values.
Coming from a C Linux development background, I'm really used to GNU
functions that both perform a particular action and then return some
diagnostic information. Some of my functions follow this pattern; the
return value is meant to be used or discarded at will. Obviously this
warning runs counter to my design philosophy, so I've disabled it.